getUserId 변경 및 swagger 관련 편의성 관련 update
