+++ src/main/java/com/website/military/controller/WordController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.website.military.domain.Entity.Word;
 import com.website.military.domain.dto.word.request.AddWordToWordSetDto;
 import com.website.military.domain.dto.word.request.ExistWordDto;
+import com.website.military.domain.dto.word.response.ExistWordResponseDto;
 import com.website.military.service.WordService;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -13,6 +14,7 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,32 +22,39 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 
 
 @RestController
 @RequestMapping("/api/word")
+@Tag(name = "Word", description = "단어관련 API")
 public class WordController {
     @Autowired
     private WordService wordService;
 
     // GET
-    @GetMapping("/exists")
-    public ResponseEntity<?> existWord(@RequestParam ExistWordDto dto, HttpServletRequest request) {
+
+    // POST
+    @Operation(summary = "exist to word", description = "단어가 있는지 체크")
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "EXIST",
+            content = {@Content(schema = @Schema(implementation = ExistWordResponseDto.class))}),
+        @ApiResponse(responseCode = "200", description = "해당하는 단어가 DB에 없습니다."),
+        @ApiResponse(responseCode = "401", description = "토큰에 해당하는 사용자가 없습니다."),
+        @ApiResponse(responseCode = "500", description = "서버 에러")
+    })
+    @PostMapping("/exists")
+    public ResponseEntity<?> existWord(@RequestBody ExistWordDto dto, HttpServletRequest request) {
         ResponseEntity<?> entity = wordService.existWord(dto, request);
         return entity;
     }
-    
 
-    // POST
     @Operation(summary = "add Word to Wordsets", description = "단어를 단어세트에 넣기.")
     @ApiResponses(value = {
         @ApiResponse(responseCode = "200", description = "성공",
             content = {@Content(schema = @Schema(implementation = Word.class))}),
-        @ApiResponse(responseCode = "400", description = "단어셋을 만든 사람과 사용하는 사용자가 다릅니다."),
-        @ApiResponse(responseCode = "400", description = "토큰에 해당하는 사용자가 없습니다."),
+        @ApiResponse(responseCode = "401", description = "단어셋을 만든 사람과 사용하는 사용자가 다릅니다."),
+        @ApiResponse(responseCode = "401", description = "토큰에 해당하는 사용자가 없습니다."),
         @ApiResponse(responseCode = "400", description = "단어셋의 입력이 잘못되었습니다."),
         @ApiResponse(responseCode = "500", description = "서버 에러")
     })
